Theme implementation
actually works now, though not finished
